solve some comments.
