+++ .env.template
@@ -164,6 +164,7 @@ KNOWLEDGE_GRAPH_EXTRACT_SEARCH_TOP_SIZE=5
 KNOWLEDGE_GRAPH_EXTRACT_SEARCH_RECALL_SCORE=0.3
 KNOWLEDGE_GRAPH_SIMILARITY_SEARCH_TOP_SIZE=5
 KNOWLEDGE_GRAPH_SIMILARITY_SEARCH_RECALL_SCORE=0.7
+KNOWLEDGE_GRAPH_TEXT_SEARCH_TOP_SIZE=5
 KNOWLEDGE_GRAPH_COMMUNITY_SEARCH_TOP_SIZE=20
 KNOWLEDGE_GRAPH_COMMUNITY_SEARCH_RECALL_SCORE=0.0
 
@@ -171,7 +172,7 @@ GRAPH_COMMUNITY_SUMMARY_ENABLED=True  # enable the graph community summary
 TRIPLET_GRAPH_ENABLED=True  # enable the graph search for triplets
 DOCUMENT_GRAPH_ENABLED=True  # enable the graph search for documents and chunks
 SIMILARITY_SEARCH_ENABLED=True # enable the similarity search for entities and chunks
-TEXT2GQL_SEARCH_ENABLED=False # enable the text2gql search for entities and relations.
+TEXT_SEARCH_ENABLED=False # enable the text search for entities and relations.
 
 KNOWLEDGE_GRAPH_CHUNK_SEARCH_TOP_SIZE=5  # the top size of knowledge graph search for chunks
 KNOWLEDGE_GRAPH_EXTRACTION_BATCH_SIZE=20  # the batch size of triplet extraction from the text
