+++ dbgpt/storage/graph_store/base.py
@@ -31,9 +31,9 @@ class GraphStoreConfig(BaseModel):
         default=False,
         description="Enable similarity search or not.",
     )
-    enable_text2gql_search: bool = Field(
+    enable_text_search: bool = Field(
         default=False,
-        description="Enable text2gql search or not.",
+        description="Enable text search or not.",
     )
 
 
@@ -46,7 +46,7 @@ class GraphStoreBase(ABC):
         self._conn = None
         self.enable_summary = config.enable_summary
         self.enable_similarity_search = config.enable_similarity_search
-        self.enable_text2gql_search = config.enable_text2gql_search
+        self.enable_text_search = config.enable_text_search
 
     @abstractmethod
     def get_config(self) -> GraphStoreConfig:
