+++ docs/static/img/chat_knowledge/graph_rag/comparison_result_for_keywords_search.png
